Move tests to correct unit
